Add int->long promotion on overflow
This commit is a straightforward implementation that doesn't include
any speculation-like optimizations.  For operations that can overflow,
just relax the type-return-specification to UNKNOWN, and do the overflow
checks in the runtime.  This means that we no longer emit fast native
integer instructions even if we know operands are ints.

Will have to add optimizations:
- range analysis so that we can know there won't be overflow
- deopt-on-overflow so that we can work with unboxed ints even if there's potential overflow

+++ src/codegen/llvm_interpreter.cpp
@@ -650,8 +650,8 @@ Box* interpretFunction(llvm::Function* f, int nargs, Box* closure, Box* generato
                                 args[0].n, args[1].d, args[2].d, args[3].n);
                             break;
                         case 0b101010:
-                            r = reinterpret_cast<int64_t (*)(double, int, double, int64_t)>(f)(args[0].d, args[1].n,
-                                                                                               args[2].d, args[3].n);
+                            r = reinterpret_cast<int64_t (*)(double, int64_t, double, int64_t)>(f)(
+                                args[0].d, args[1].n, args[2].d, args[3].n);
                             break;
                         case 0b1000000: // 64
                             r = reinterpret_cast<int64_t (*)(int64_t, int64_t, int64_t, int64_t, int64_t)>(f)(
